# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickShare API - share text and files between devices on the same network
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: string;
  title: string;
  excerpt: string;
  content: string;
  imageDataUrl: string | null;
  publishedAt: Date;
}