# lib/api-zod/src/generated/types/board.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickShare API - share text and files between devices on the same network
 * OpenAPI spec version: 0.1.0
 */
import type { SharedFile } from "./sharedFile";
import type { TextSnippet } from "./textSnippet";

export interface Board {
  /** Identifier for the network room (derived from request IP) */
  roomId: string;
  text?: TextSnippet | null;
  files: SharedFile[];
  deviceCount: number;
  /** Minutes until current text snippet auto-expires */
  expiresInMinutes: number;
}