# lib/api-zod/src/generated/types/sharedFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickShare API - share text and files between devices on the same network
 * OpenAPI spec version: 0.1.0
 */

export interface SharedFile {
  id: string;
  name: string;
  sizeBytes: number;
  mimeType: string;
  createdAt: Date;
  expiresAt: Date;
  deviceLabel: string;
}