# lib/api-zod/src/generated/types/uploadFileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QuickShare API - share text and files between devices on the same network
 * OpenAPI spec version: 0.1.0
 */

export interface UploadFileBody {
  /**
   * @minLength 1
   * @maxLength 255
   */
  name: string;
  mimeType: string;
  /** Base64-encoded file content */
  dataBase64: string;
}